removed update assets component, removed asset from being saved in local storage in anticipation of backend route
