+++ apps/frontend/src/app/views/assets/AssetList.tsx
@@ -1,18 +1,23 @@
 // import { useContext } from 'react';
 // import { AssetContext } from './AssetContext';
 // import { AssetContext } from './Index';
-import { UpdateAssets } from './UpdateAssets';
+// import { UpdateAssets } from './UpdateAssets';
 
 const AssetList = (props: any) => {
 //   const { assets } = useContext(AssetContext);
 
   return (
-    <ul>
-      {props.asset.map((asset: any) => (
-        <UpdateAssets key={props.asset.id} asset={props.asset.asset} />
-      ))}
-    </ul>
-  );
+  //   <ul>
+  //     {props.asset.map((asset: any) => (
+  //       // <UpdateAssets key={props.asset.id} asset={props.asset.asset} />
+  //     ))}
+  //   </ul>
+  // );
+
+  <div>
+    <p>assets</p>
+  </div>
+  )
 };
 
 export default AssetList;
